🐛Fix : 모든 브랜치 병합 후 1차 버그 수정
