+++ .gitignore
@@ -41,3 +41,4 @@ next-env.d.ts
 
 # package-lock.json
 .package-lock.json
+/package-lock.json
