--- src/pages/api/auth/[...nextauth].ts
@@ -1,19 +1,0 @@
-import NextAuth from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-
-export default NextAuth({
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-  ],
-  callbacks: {
-    async session({ session, token, user }) {
-      return session;
-    },
-    async signIn({ user, account, profile, email, credentials }) {
-      return true;
-    },
-  },
-});
