+++ src/pages/index.tsx
@@ -1,9 +1,5 @@
-import Modal from '@/components/Modal';
-import { useEffect, useState } from 'react';
-import { signIn, signOut, useSession } from 'next-auth/react';
-import axios from 'axios';
-import Cookies from 'js-cookie';
+import Randing from './landing';
 
 export default function Home() {
-  return <></>;
+  return <Randing />;
 }
