+++ src/pages/privacy-policy/style.module.scss
