--- src/pages/test-page/index.tsx
@@ -1,14 +1,0 @@
-import Card from '@/components/Card';
-export default function page() {
-  return (
-    <>
-      <h1>Hello, world!</h1>
-      <p>This is a Next.js app with TypeScript, ESLint, Prettier, and SCSS.</p>
-      <Card
-        updatedAt={'Wed Jun 26 2024 17:48:18 GMT+0900'}
-        content={'Lorem ipsum dolor sit amet consectetur. Metus amet habitant nunc consequat. Tldkd'}
-        createdAt={'Wed Jun 26 2024 17:48:18 GMT+0900'}
-      />
-    </>
-  );
-}
