+++ src/styles/index.scss
@@ -1,4 +1,3 @@
 @import './color.scss';
 @import './font.scss';
 @import './variable.scss';
-// @import './normalize.css';
