+++ src/styles/normalize.css
@@ -1,5 +1,3 @@
-@import './index.scss';
-
 .droppable {
   background-color: #f8f9fa;
   padding: 16px;
